version: v2.6.1版本发布 【新增】一句话生成视频新增 Sora-2-Pro 模型支持 【新增】小程序一句话生成视频支持 15 秒时长 生成 【优化】抖音发布增加话题数量限制，避免超过 5 个话题导致发布失败 【优化】AI 智剪字幕效果恢复为原版表现 【优化】小程序移除各模块登录验证流程，提升使用流畅度 【缺陷修复】Sora 视频合成失败后，后台详情页仍显示“生成中”的状态未更新问题 【缺陷修复】一句话生成视频请求超时后重复请求，首次超时请求未退还扣费的问题 【缺陷修复】混剪任务创建时算力消耗计算错误的问题 【缺陷修复】后台混剪创建记录未显示失败状态的问题 【缺陷修复】拼图子任务已完成，但主任务仍显示生成中的问题 【缺陷修复】数字人标准版无法选择原视频音色的问题 【缺陷修复】数字人口播文案扣费异常为 0 的问题 【缺陷修复】素材混剪中，无法通过录音或聊天音频进行音色克隆的问题（PC / iOS / Android 不同步） 【缺陷修复】小程序数字人克隆音色无法录音的问题 【缺陷修复】小程序数字人创建 AI 文案，选择生成长文案后点击使用不生效的问题 【缺陷修复】PC、小程序与手机端发布状态不同步的问题 【缺陷修复】发布过程中 AI 自动生成标题或正文时错误提示“扣费问题”的异常 【缺陷修复】直接选择智能文案时，标题与正文错位添加的问题 【缺陷修复】向量知识库在通用聊天中最低相似度配置未生效的问题 【缺陷修复】获客线索手机号无法正常获取的问题 【缺陷修复】删除自动加好友任务时，导致同一主任务下其他子任务无法删除的问题 【缺陷修复】小程序智能拼图上传素材提示错误的问题

+++ server/app/api/logic/device/TaskLogic.php
@@ -313,7 +313,10 @@ class TaskLogic extends ApiLogic
                     if (!$taskinfo) {
                         throw new \Exception('自动加好友任务不存在');
                     }
-                    SvCrawlingManualTask::where('id', $params['sub_task_id'])->select()->delete();
+                    $count = SvDeviceTask::where('sub_task_id', $params['sub_task_id'])->where('task_type',DeviceEnum::TASK_SOURCE_FRIENDS)->where('user_id', self::$uid)->count();
+                    if ($count == 1) {
+                        SvCrawlingManualTask::where('id', $params['sub_task_id'])->select()->delete();
+                    }
                     break;
 
                 case DeviceEnum::TASK_SOURCE_CLUES:
